+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ## Unreleased
 
+* Add new chart component options ([PR #4318](https://github.com/alphagov/govuk_publishing_components/pull/4318))
 * Add shared helper and component wrapper helper to govspeak component ([PR #4325](https://github.com/alphagov/govuk_publishing_components/pull/4325))
 
 ## 44.4.2